+++ webhook_server/tests/test_add_reviewer_action.py
@@ -1,5 +1,7 @@
 import logging
 
+from webhook_server.libs.issue_comment_handler import IssueCommentHandler
+
 
 class User:
     def __init__(self, username):
@@ -28,7 +30,8 @@ class PullRequest:
 def test_add_reviewer_by_user_comment(caplog, process_github_webhook):
     process_github_webhook.repository = Repository()
     process_github_webhook.pull_request = PullRequest()
-    process_github_webhook._add_reviewer_by_user_comment("user1")
+    issue_comment_handler = IssueCommentHandler(github_webhook=process_github_webhook)
+    issue_comment_handler._add_reviewer_by_user_comment("user1")
     caplog.set_level(logging.DEBUG)
     assert "Adding reviewer user1 by user comment" in caplog.text
 
@@ -36,6 +39,7 @@ def test_add_reviewer_by_user_comment(caplog, process_github_webhook):
 def test_add_reviewer_by_user_comment_invalid_user(caplog, process_github_webhook):
     process_github_webhook.repository = Repository()
     process_github_webhook.pull_request = PullRequest()
-    process_github_webhook._add_reviewer_by_user_comment("user2")
+    issue_comment_handler = IssueCommentHandler(github_webhook=process_github_webhook)
+    issue_comment_handler._add_reviewer_by_user_comment("user2")
     caplog.set_level(logging.DEBUG)
     assert "not adding reviewer user2 by user comment, user2 is not part of contributers" in caplog.text
